+++ Changelog.md
@@ -1,5 +1,10 @@
 # CodeImatic.share Changelog
 
+18/06/2023
+
+* Start of new  ProjectTemplate class
+
+
 18/04/2023
 
 * Now supports Delphi 10.4 and Delphi 11 NovuscodeLibrary v0.1.0